pyproject.toml
[tool.poetry]
name = "speedtest"
version = "0.1.0"
description = ""
authors = ["Вадим Акимов <sfc.coltrane@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
djangorestframework = "^3.14"
ruff = "^0.2.2"
django-redis = "^5.4.0"
plotly = "^5.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
